cydbank_web/requirements.txt
aioredis==1.3.1
asgi-redis==1.4.3
asgiref==3.2.7
astroid==2.4.1
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
beautifulsoup4==4.9.0
certifi==2020.4.5.1
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
colorama==0.4.3
confusable-homoglyphs==3.2.0
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.4
django-appconf==1.0.4
django-bootstrap4==1.1.1
django-ckeditor==5.9.0
django-crispy-forms==1.9.0
django-filter==2.2.0
django-import-export==2.0.2
django-js-asset==1.2.2
django-mathfilters==1.0.0
django-registration==3.1
django-tables2==2.2.1
et-xmlfile==1.0.1
haikunator==2.1.0
hiredis==1.0.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
isort==4.3.21
jdcal==1.4.1
lazy-object-proxy==1.4.3
MarkupPy==1.14
mccabe==0.6.1
msgpack==0.6.2
msgpack-python==0.5.6
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pylint==2.5.2
pyOpenSSL==19.1.0
pypiwin32==223
PySocks==1.7.1
python-dotenv==0.13.0
pytz==2019.3
pywin32==227
PyYAML==5.3.1
redis==2.10.6
requests==2.23.0
service-identity==18.1.0
six==1.15.0
soupsieve==2.0
sqlparse==0.3.1
tablib==1.1.0
toml==0.10.1
twilio==6.10.1
Twisted==20.3.0
txaio==20.4.1
typed-ast==1.4.1
urllib3==1.25.9
websocket-client==0.57.0
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
zope.interface==5.1.0
